Moved CryptoNoteConfig.h to lib/Global
