+++ lib/CMakeLists.txt
@@ -74,9 +74,14 @@ set(QwertycoinFramework_BlockchainExplorer_SOURCES
     "${CMAKE_CURRENT_LIST_DIR}/BlockchainExplorer/IBlockchainObserver.h"
 )
 
+set(QwertycoinFramework_BlockchainExplorer_LIBS
+    QwertycoinFramework::Global
+)
+
 add_library(QwertycoinFramework_BlockchainExplorer STATIC ${QwertycoinFramework_BlockchainExplorer_SOURCES})
 add_library(QwertycoinFramework::BlockchainExplorer ALIAS QwertycoinFramework_BlockchainExplorer)
 target_include_directories(QwertycoinFramework_BlockchainExplorer PRIVATE ${QwertycoinFramework_INCLUDE_DIRS})
+target_link_libraries(QwertycoinFramework_BlockchainExplorer PRIVATE ${QwertycoinFramework_BlockchainExplorer_LIBS})
 
 # QwertycoinFramework::Common
 
@@ -142,6 +147,7 @@ set(QwertycoinFramework_Common_SOURCES
 set(QwertycoinFramework_Common_LIBS
     Threads::Threads
     QwertycoinFramework::Crypto
+    QwertycoinFramework::Global
 )
 
 add_library(QwertycoinFramework_Common STATIC ${QwertycoinFramework_Common_SOURCES})
@@ -320,6 +326,7 @@ target_link_libraries(QwertycoinFramework_CryptoNoteProtocol PRIVATE ${Qwertycoi
 set(QwertycoinFramework_Global_SOURCES
     "${CMAKE_CURRENT_LIST_DIR}/Global/Constants.cpp"
     "${CMAKE_CURRENT_LIST_DIR}/Global/Constants.h"
+    "${CMAKE_CURRENT_LIST_DIR}/Global/CryptoNoteConfig.h"
 )
 
 add_library(QwertycoinFramework_Global STATIC ${QwertycoinFramework_Global_SOURCES})
@@ -478,6 +485,7 @@ set(QwertycoinFramework_P2p_SOURCES
 set(QwertycoinFramework_P2p_LIBS
     MiniUPnP::miniupnpc
     QwertycoinFramework::CryptoNoteProtocol
+    QwertycoinFramework::Global
 )
 
 if(WIN32 AND MSVC)
