+++ lib/Global/CryptoNoteConfig.h
