+++ lib/P2p/NetNodeCommon.h
@@ -21,9 +21,9 @@
 
 #pragma once
 
-#include <P2p/P2pProtocolTypes.h>
-#include <../src/config/CryptoNoteConfig.h>
 #include <CryptoNote.h>
+#include <Global/CryptoNoteConfig.h>
+#include <P2p/P2pProtocolTypes.h>
 
 namespace CryptoNote {
